// src/app/models/pensionado.ts
export class Pensionado {
    iCodigoTipoElegibilidad: number;
    vcDescripcion: string;
    iRUN: number;
    cDV: string;
    iRUTEntidad: number;
    cDVEntidad: string;
    vcApellidoPaterno: string;
    vcApellidoMaterno: string;
    vcNombre: string;
    dFechaNacimiento: string;
    iEdad: number;
    dFechaDefuncion: string;
    cSexo: string;
    dFechaPension: string;
    iNumeroCotizacionesPACS: number;
    iCotizacionesAFP: number;
    iCotizacionesDIPRECA: number;
    iCotizacionesCAPREDENA: number;
    iCotizacionesIPS: number;
    cBonoReconocimientoCAPREDENA: string;
    cBonoReconocimientoDIPRECA: string;
    cBonoReconocimientoIPS: string;
    nMontoPensionPagadaRP: number;
    nMontoPensionPegadaRV: number;
    nPensionDl3500: number;
    nMontoPensionDIPRECA: number;
    nMontoPensionDIPRECAUF: number;
    nMontoPensionCAPREDENA: number;
    nMontoPensionCAPREDENAUF: number;
    nMontoPensionMutual: number;
    nMontoPensionMutualUF: number;
    nMontoPilarSolidario: number;
    nMontoPilarSolidarioUF: number;
    nMontoSubsidioDependencia: number;
    nMontoSubsidioDependenciaUF: number;
    nMontoSeguroDependecia: number;
    nMontoSeguroDependeciaUF: number;
    nMontoPensionIPS: number;
    nMontoPensionIPSUF: number;
    nMontoTGR: number;
    nMontoTGRUF: number;
    nTotalOtrosP: number;
    nTotalOtrosUF: number;
    dFechaDevengamiento: string;
    vcEntidadPagadora: string;
    nMontoBeneficioFijopesos: number;
    nMontoBeneficioVariablePesos: number;
    nMontoREBAJABeneficioFijopesos: number;
    nMontoREBAJABeneficioVariablePesos: number;
    nValorUFUtilizado: number;
    cPeriodo: string;
    dFechaCreacion: string;
    dFechaActualizacion: string;
    nMontoInicialPACS: number;
    nMontoInicialPACSUF: number;
    nPensionTotalUF: number;
    nTotalPensiones: number;
    vcNroCotizaciones: string;
    vcDl3500: string;
    vcCumpleEdad: string;
    vcPacVigente: string;
    vcCotizacionPac: string;
    vcAntesLey: string;
    vcCumpleFijo: string;
    vcCumpleVariable: string;
    vcPensionTotalUF: string;
    vcTotalPensiones: string;
    vcMensaje: string;
}